==> 1_sample_labelling.ipynb <==
import pandas as pd
import numpy as np
from collections import Counter
from tqdm import tqdm

species="human"
gsm4sig_version = 2

bulk_study_meta = pd.read_csv("data/"+species+"_bulk_study_meta.csv", index_col=0)
frequent_terms = pd.read_csv(f"data/{species}v{gsm4sig_version}_frequent_terms.csv", index_col=0)

frequent_terms
# Category 0: Neutral
# Category 1: Control term
# Category 2: Perturbation term

bulk_study_meta["sample_title_set"] = (bulk_study_meta["sample_title"]             # processing the sample titles by 
                                       .str.lower().str.findall(r"[a-z]+")         # extracting the keywords and 
                                       .apply(frozenset))                          # saving them as a (frozen) set

ctrl_set = set(frequent_terms.loc[frequent_terms["Category"] == 1]["Label"])
pert_set = set(frequent_terms.loc[frequent_terms["Category"] == 2]["Label"])

bulk_study_meta

def categorize_cp(test_set, ctrl_set = ctrl_set, pert_set = pert_set): # categorizes sample as ctrl or pert
    difference = len(test_set & ctrl_set) - len(test_set & pert_set)   # 0 - undetermined
    if difference == 0: return 0                                       # 1 - control group
    return 1 if difference > 0 else 2                                  # 2 - perturbation group

bulk_study_meta["cp_group"] = bulk_study_meta["sample_title_set"].apply(categorize_cp)

len(bulk_study_meta["series_id"].unique()) # unique studies before filtering

no_undet = []
for gse in bulk_study_meta.groupby("series_id")["sample_title_set"]:   
    if ((gse[1].value_counts() >= 2).all()): no_undet.append(gse[0])  # keeping studies where theres >= 2 samples per category
        
no_undet_study = bulk_study_meta[bulk_study_meta["series_id"].isin(no_undet)].copy()

len(no_undet)                            # unique studies post filtering

#gse_per_term = []                                                                              # counts unique gse for each 
#def det_gse(test_set, term): return 1 if len(test_set & {term}) > 0 else 0                     # control or perturbation term
#for i in frequent_terms["Label"]:
#    bulk_study_meta["det_gse"] = bulk_study_meta["sample_title_set"].apply(det_gse, term=i)
#    gse_per_term.append(len(bulk_study_meta.loc[bulk_study_meta["det_gse"] == 1]["series_id"].unique()))
#    bulk_study_meta["det_gse"] = np.zeros(162372)
    
#frequent_terms["num_gse"] = gse_per_term
#frequent_terms.to_csv(species+"_frequent_terms_gse_count.csv", index=False)

#def gsm_title_cluster(series, cluster_esp = 0.01, cluster_min_samp = 2):                     # grouping gsm by clustering words
#    mlb = MultiLabelBinarizer()
#    term_mat = mlb.fit_transform(series["sample_title_set"])
#    if 1 in term_mat.shape: return np.zeros(len(series)).astype(int)
#    pca = PCA()
#    pca.fit(term_mat.T)
#    clustering = DBSCAN(eps=0.01, min_samples=2)
#    #print(list(zip(pca.components_[0], pca.components_[1])))
#    #clustering.fit(list(zip(pca.components_[0], pca.components_[1])))
#    clustering.fit(pca.components_[0].reshape(-1, 1))
#    return clustering.labels_


# for gse in no_undet_study["series_id"].unique():
#    series = no_undet_study.loc[no_undet_study["series_id"] == gse]
#    no_undet_study.loc[no_undet_study["series_id"] == gse, "gsm_cluster"] = gsm_title_cluster(series)

def count_series(series):
    '''
    Categorize an input series and return (undet, ctrl, pert) tuple
    '''
    series_counter = (Counter(list(zip(*series.groupby(["cp_group", "sample_title_set"])
                                       .groups.keys()))[0]))
                                          # Subgroup counts for:
    return (min(series_counter[0], 3),    # undetermined group
            min(series_counter[1], 3),    # control group
            min(series_counter[2], 3))    # perturbation group

gse_types = {}
for n, gse in tqdm(enumerate(no_undet_study["series_id"].unique())):               # categorizing each of the filtered study
    series = no_undet_study.loc[no_undet_study["series_id"] == gse]
    gse_types[gse] = count_series(series)
    
categorized_gse = pd.DataFrame({"gse":gse_types.keys(), "gse_type":gse_types.values()})

print("(?, c, p)    freq   3 := >2")
print(categorized_gse["gse_type"].value_counts().to_string())

def extrap_score(study_cat):
    '''
    Determine extrapolation score for a given study category tuple by 
    penalizing each of inferred type (ctrl, pert) and direction
    '''
    undet, ctrl, pert = study_cat
    score = 0
    if undet == 0 or pert == 0:                 #(0, x, y) or (y, x, 0)
        if ctrl == 1:                           #(0, 1, y) or (y, 1, 0)
            if undet != 0: score += 1           #penalty for inferred pert
        if ctrl < 1:                            #(0, 0, y) or (y, 0, 0)
            score += 1                          #penalty for inferred ctrl
            score += 2 if undet != 0 else 1     #penalty for inferred direction and/or inferred pert
    else:
        if ctrl < 1:                            #(x, 0, y)
            score += 1 if undet == 1 else -1    #penality for inferred ctrl
        else:
            score = -1                          #unaccounted for
    return score
        

def to_calc_sig_format(study_cat, studies = no_undet_study, categorized_gse = categorized_gse):
    '''
    Convert studies corresponding to an input study category type to 
    a signature-calculation-friendly dataframe format with rows: 
    gse, [ctrl_gsm0, ctrl_gsm1, ...], [pert_gsm0, pert_gsm1, ...], extrap_score
    '''
    undet, ctrl, pert = study_cat
    study_group = studies.loc[studies.series_id.isin(
        categorized_gse.loc[categorized_gse["gse_type"] == study_cat]["gse"])].copy()
    if ctrl == 0 and undet == 1: #(1, 0, x)
        study_group.replace({'cp_group':{0:"ctrl_gsm", 2:"pert_gsm"}}, inplace=True)
    else:
        study_group.replace({'cp_group':{1:"ctrl_gsm", 0:"pert_gsm", 2:"pert_gsm"}}, inplace=True)
        
    if ctrl == 0 and (undet == 0 or pert == 0):
        study_cp_series = study_group.groupby(["series_id", "sample_title_set"])["cp_group"].unique()
        new_ctrl_idx = study_cp_series.index[::2]
        study_cp_series[study_cp_series.index.isin(new_ctrl_idx)] = "ctrl_gsm"
        study_group_map = dict(study_cp_series.explode())
        study_group = (study_group.assign(cp_group=study_group[["series_id","sample_title_set"]]
                                         .apply(tuple, axis=1).map(study_group_map)))
            
    return (study_group
            .groupby(["series_id", "cp_group","sample_title_set"])["geo_accession"]
            .aggregate(lambda x: list(x))                  # aggregates gsms as a list (ie [gsm1, gsm2, ...])
           #.aggregate(lambda x: "|".join(x))              # aggregates gsms separated by pipes (ie gsm1|gsm2|...)
            .unstack()
            .apply(lambda x: list(x.dropna()), axis=1)
            .unstack()
            .explode("ctrl_gsm")
            .explode("pert_gsm")
            .assign(extrap_score=extrap_score(study_cat)))

# valid types used to generate final list of gsms for signature calculation
valid_types = [(0, 1, 1), (0, 1, 2), (0, 1, 3), 
               (1, 1, 0), (2, 1, 0), (3, 1, 0), 
               (0, 0, 2), (2, 0, 0),
               (1, 0, 1), (1, 0, 2), (1, 0, 3)]
gsm4sig = pd.concat([to_calc_sig_format(i) for i in valid_types]).reset_index()
gsm4sig["series_id"] = gsm4sig["series_id"].replace(regex=r"\t", value="-")
gsm4sig.to_csv(f"data/{species}_gsm4sig_v{gsm4sig_version}.csv")
