==> 4_score_extraction.ipynb <==
import pandas as pd
import numpy as np

species = "human"
gsm4sig_version = 2

gsm4sig = pd.read_csv(f"data/{species}_gsm4sig_v{gsm4sig_version}.csv", index_col=0)
gsm4sig["sig_num"] = gsm4sig.groupby("series_id").cumcount()+1
fc_data = pd.read_csv(f"data/all_{species}_fc.csv", index_col=0, nrows = 2)
gse_trunc = gsm4sig["series_id"].str.split("-").map(lambda x: x[0])

gsm4sig

fc_data

extrap_score = pd.DataFrame([np.repeat(-1, len(fc_data.columns))], columns = fc_data.columns, index = ['score'])

for study in fc_data.columns:
    gse, sig_num = study.split()[-1].split("_")
    try:
        extrap_score.at['score', study] = (gsm4sig.loc[np.isin(gse_trunc, gse) &
                                                       (gsm4sig.sig_num == int(sig_num)), 
                                                       "extrap_score"].item())
    except:
        print(study) #some studies had their sub/super series order flipped, extrap_score manually inputted

extrap_score.loc["score"].value_counts()

extrap_score.to_csv(f"all_{species}_score.csv")
